test:(lighthouse): add lighthouse tests
