+++ package.json
@@ -59,8 +59,10 @@
     "jest-circus": "^26.6.3",
     "jest-dev-server": "^4.4.0",
     "jest-environment-puppeteer": "^4.4.0",
+    "jest-expect-message": "^1.0.2",
     "jest-image-snapshot": "^4.3.0",
     "jest-puppeteer": "^4.4.0",
+    "lighthouse": "^7.0.0",
     "postcss-import": "^12.0.1",
     "postcss-media-variables": "^2.0.1",
     "postcss-nested": "^4.2.3",
