+++ tests/setup/config.js
@@ -1,6 +1,16 @@
+const parser = require("fast-xml-parser");
+const fs = require("fs");
 const puppeteer = require("puppeteer");
 
+const sitemap = fs.readFileSync(`${process.cwd()}/dist/sitemap.xml`, "utf-8");
+const sitemapJson = parser.parse(sitemap);
+const pages = sitemapJson.urlset.url.map(({ loc }) => {
+	const url = new URL(loc);
+	return url.pathname;
+});
+
 module.exports = {
+	pages,
 	browserTimeout: 30 * 1000,
 	port: 3001,
 	devices: {
