+++ tests/setup/setup-files-after-env.js
@@ -1,3 +1,4 @@
+require("jest-expect-message");
 const { toMatchImageSnapshot } = require("jest-image-snapshot");
 
 expect.extend({ toMatchImageSnapshot });
